create area hover/click manager
