+++ src/components/AreaHighlightManager/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import AreaHighlightManagerComponent from './area-highlight-manager-component';
+
+function AreaHighlightManager(props) {
+  return <AreaHighlightManagerComponent {...props} />;
+}
+
+export default AreaHighlightManager;
